Assignment/requirements.txt
fastapi
uvicorn
jinja2
mcp
pywinauto
google-generativeai
